delete deprecated loader code
